refactor : [017-trip-invite] 여행 기록장 참여자 조회 기능 Participant 컨트롤러 및 서비스로 분리
