+++ src/main/java/com/trip/diary/service/ParticipantService.java
@@ -1,0 +1,45 @@
+package com.trip.diary.service;
+
+import com.trip.diary.domain.model.Member;
+import com.trip.diary.domain.model.Participant;
+import com.trip.diary.domain.model.Trip;
+import com.trip.diary.domain.repository.ParticipantRepository;
+import com.trip.diary.domain.repository.TripRepository;
+import com.trip.diary.dto.ParticipantDto;
+import com.trip.diary.exception.ErrorCode;
+import com.trip.diary.exception.TripException;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+import static com.trip.diary.exception.ErrorCode.NOT_FOUND_TRIP;
+
+@Service
+@RequiredArgsConstructor
+public class ParticipantService {
+
+    private final TripRepository tripRepository;
+
+    private final ParticipantRepository participantRepository;
+
+    @Transactional
+    public List<ParticipantDto> getTripParticipants(Long tripId, Member member) {
+        Trip trip = tripRepository.findById(tripId)
+                .orElseThrow(() -> new TripException(NOT_FOUND_TRIP));
+
+        List<Participant> participants = trip.getParticipants();
+
+        if (trip.isPrivate() && !participantRepository.existsByTripAndMember(trip, member)) {
+            throw new TripException(ErrorCode.NOT_AUTHORITY_READ_TRIP);
+        }
+
+        return participants
+                .stream()
+                .map(participant ->
+                        ParticipantDto.of(participant, member.getId()))
+                .collect(Collectors.toList());
+    }
+}
