+++ src/test/java/com/trip/diary/service/TripServiceTest.java
@@ -240,164 +240,6 @@ class TripServiceTest {
     }
 
     @Test
-    @DisplayName("여행 기록장 참여 멤버 조회 성공 - 조회자가 리더인 경우")
-    void getTripParticipantsTest_successWhenReaderIsTripLeader() {
-        //given
-        Trip trip = Trip.builder()
-                .id(1L)
-                .title("임의의 타이틀")
-                .isPrivate(true)
-                .description("임의의 설명")
-                .leader(member)
-                .participants(List.of(
-                        Participant.builder()
-                                .member(member)
-                                .type(ParticipantType.ACCEPTED)
-                                .build(),
-                        Participant.builder()
-                                .member(participant1)
-                                .type(ParticipantType.PENDING)
-                                .build(),
-                        Participant.builder()
-                                .member(participant2)
-                                .type(ParticipantType.PENDING)
-                                .build()))
-                .build();
-        given(tripRepository.findById(anyLong())).willReturn(Optional.of(trip));
-        given(participantRepository.existsByTripAndMember(any(), any())).willReturn(true);
-        //when
-        List<ParticipantDto> result = tripService.getTripParticipants(1L, member);
-        //then
-        assertTrue(result.get(0).isReader());
-        assertTrue(result.get(0).isAccepted());
-        assertEquals(participant1.getNickname(), result.get(1).getNickname());
-        assertFalse(result.get(1).isReader());
-        assertFalse(result.get(1).isAccepted());
-    }
-
-    @Test
-    @DisplayName("여행 기록장 참여 멤버 조회 성공 - 조회자가 초대받은 사람인 경우")
-    void getTripParticipantsTest_successWhenReaderIsParticipants() {
-        //given
-        Trip trip = Trip.builder()
-                .id(1L)
-                .title("임의의 타이틀")
-                .isPrivate(true)
-                .description("임의의 설명")
-                .leader(member)
-                .participants(List.of(
-                        Participant.builder()
-                                .member(member)
-                                .type(ParticipantType.ACCEPTED)
-                                .build(),
-                        Participant.builder()
-                                .member(participant1)
-                                .type(ParticipantType.PENDING)
-                                .build(),
-                        Participant.builder()
-                                .member(participant2)
-                                .type(ParticipantType.PENDING)
-                                .build()))
-                .build();
-        given(tripRepository.findById(anyLong())).willReturn(Optional.of(trip));
-        given(participantRepository.existsByTripAndMember(any(), any())).willReturn(true);
-        //when
-        List<ParticipantDto> result = tripService.getTripParticipants(1L, participant1);
-        //then
-        assertFalse(result.get(0).isReader());
-        assertTrue(result.get(0).isAccepted());
-        assertNotEquals(participant1.getNickname(), result.get(1).getNickname());
-        assertTrue(result.get(1).isReader());
-        assertFalse(result.get(1).isAccepted());
-    }
-
-    @Test
-    @DisplayName("여행 기록장 참여 멤버 조회 성공 - 조회자가 참여자가 아니지만 해당 여행기록장이 모두 공개임")
-    void getTripParticipantsTest_successWhenTripIsPublic() {
-        //given
-        Member leader = Member.builder()
-                .id(4L)
-                .username("1234qwert")
-                .nickname("오땡땡")
-                .password("1234567")
-                .profilePath(null)
-                .phone("01011111114")
-                .build();
-
-        Trip trip = Trip.builder()
-                .id(1L)
-                .title("임의의 타이틀")
-                .isPrivate(false)
-                .description("임의의 설명")
-                .leader(leader)
-                .participants(List.of(
-                        Participant.builder()
-                                .member(leader)
-                                .type(ParticipantType.ACCEPTED)
-                                .build(),
-                        Participant.builder()
-                                .member(participant1)
-                                .type(ParticipantType.PENDING)
-                                .build(),
-                        Participant.builder()
-                                .member(participant2)
-                                .type(ParticipantType.PENDING)
-                                .build()))
-                .build();
-        given(tripRepository.findById(anyLong())).willReturn(Optional.of(trip));
-        //when
-        List<ParticipantDto> result = tripService.getTripParticipants(1L, member);
-        //then
-        assertFalse(result.get(0).isReader());
-        assertTrue(result.get(0).isAccepted());
-        assertEquals(participant1.getNickname(), result.get(1).getNickname());
-        assertFalse(result.get(1).isReader());
-        assertFalse(result.get(1).isAccepted());
-    }
-
-    @Test
-    @DisplayName("여행 기록장 참여 멤버 조회 실패 - 조회자가 참여자가 아니고 해당 여행 기록장이 비밀임")
-    void getTripParticipantsTest_failWhenNotAuthorityReadTrip() {
-        //given
-        Member leader = Member.builder()
-                .id(4L)
-                .username("1234qwert")
-                .nickname("오땡땡")
-                .password("1234567")
-                .profilePath(null)
-                .phone("01011111114")
-                .build();
-
-        Trip trip = Trip.builder()
-                .id(1L)
-                .title("임의의 타이틀")
-                .isPrivate(true)
-                .description("임의의 설명")
-                .leader(leader)
-                .participants(List.of(
-                        Participant.builder()
-                                .member(leader)
-                                .type(ParticipantType.ACCEPTED)
-                                .build(),
-                        Participant.builder()
-                                .member(participant1)
-                                .type(ParticipantType.PENDING)
-                                .build(),
-                        Participant.builder()
-                                .member(participant2)
-                                .type(ParticipantType.PENDING)
-                                .build()))
-                .build();
-        given(tripRepository.findById(anyLong())).willReturn(Optional.of(trip));
-        given(participantRepository.existsByTripAndMember(any(), any())).willReturn(false);
-        //when
-        TripException exception = assertThrows(TripException.class,
-                () -> tripService.getTripParticipants(1L, member));
-        //then
-        assertEquals(exception.getErrorCode(), ErrorCode.NOT_AUTHORITY_READ_TRIP);
-    }
-
-    @Test
     @DisplayName("여행 기록장에 초대할 멤버 검색 성공")
     void searchAddableMembersTest_success() {
         //given
